test(all): fold detachable-exit ordering test into shutdown_watchdog_spec
Fix spec naming: the prior spec/lib/main/detachable_shutdown_ordering_spec.rb
had no corresponding lib/main/*.rb file and described a bare string, breaking
the one-spec-per-lib-file / describe-the-real-module convention used across
spec/lib/main and spec/lib/common. lib/main/main.rb (where the wiring lives) is
a top-level script and not unit-loadable, and lib/main/detachable_client_target
covers flag parsing, not the exit path.

The test's real subject is ShutdownWatchdog's arming ordering relative to the
inline drain, so move it into spec/lib/common/shutdown_watchdog_spec.rb under
Lich::Common::ShutdownWatchdog as an integration describe. Behaviour and
assertions are unchanged; rubocop clean; 30 examples pass.
